Rename /event to /events
