+++ front/lib/error.ts
@@ -37,6 +37,7 @@ export type APIErrorType =
   | "chat_session_auth_error"
   | "template_not_found"
   | "chat_message_not_found"
+  | "event_schema_not_found"
   | "extracted_event_not_found"
   | "extracted_event_auth_error";
