+++ front/lib/swr.ts
@@ -14,7 +14,7 @@ import { GetMembersResponseBody } from "@app/pages/api/w/[wId]/members";
 import { GetProvidersResponseBody } from "@app/pages/api/w/[wId]/providers";
 import { GetChatSessionsResponseBody } from "@app/pages/api/w/[wId]/use/chats";
 import { GetEventSchemasResponseBody } from "@app/pages/api/w/[wId]/use/extract";
-import { GetExtractedEventsResponseBody } from "@app/pages/api/w/[wId]/use/extract/[marker]/event/[eId]";
+import { GetExtractedEventsResponseBody } from "@app/pages/api/w/[wId]/use/extract/[marker]/events/[eId]";
 import { AppType } from "@app/types/app";
 import { DataSourceType } from "@app/types/data_source";
 import { RunRunType } from "@app/types/run";
@@ -279,7 +279,7 @@ export function useExtractedEvents(owner: WorkspaceType, marker: string) {
     fetcher;
 
   const { data, error } = useSWR(
-    `/api/w/${owner.sId}/use/extract/${marker}/event`,
+    `/api/w/${owner.sId}/use/extract/${marker}/events`,
     extractedEventFetcher
   );
